feat(notation): allow decimal beat as base of ±n offset
A `±n` note-value offset previously bound only to an integer grid beat:
`1|1+n/4` parsed but `1|1.5+n/4` errored. Widen the offset base to an
integer-or-decimal grid beat across all four parse sites kept in parity:
both Peggy `beatValue` rules (barbeat + transform) and the two regex
mirrors in barbeat-time.ts (barBeatToMusicalBeats + parseBeatValue).

Purely additive — `oneOrMoreDecimal` subsumes integers, so no existing
parse changes and the serializer (which always emits canonical
integer+n<frac>) is untouched. Composes on every offset surface:
position, repeat-start, non-x/4 meters, and transform timeRange start +
end bounds (including the half-open `-<` marker).

Add decimal-base rows to note-value-grammar-parity, barbeat parser, and
transform timeRange tests; update BarBeat/Transforms specs + the skill
clause. Extract the transform "time range selectors" tests into their own
file to stay under the per-function line limit.
