+++ src/notation/barbeat/parser/tests/barbeat-parser-time.test.ts
@@ -60,6 +60,21 @@ describe("BarBeatScript Parser - time declarations", () => {
     ]);
   });
 
+  it("parses decimal-base note-value offsets (1|1.5+n/4)", () => {
+    // The offset base may be a decimal beat, not just an integer: `1.5+n/4` =
+    // beat 1.5 + a quarter note; `2.25-n/24` = beat 2.25 − an eighth triplet.
+    expect(
+      parser.parse("1|1.5+n/4 C3 1|2.25-n/24 D3 1|1.5+n/8 E3"),
+    ).toStrictEqual([
+      { bar: 1, beat: obeat(1.5, 1, 4) },
+      { pitch: 60 },
+      { bar: 1, beat: 2.25 - (1 / 24) * 4 },
+      { pitch: 62 },
+      { bar: 1, beat: obeat(1.5, 1, 8) },
+      { pitch: 64 },
+    ]);
+  });
+
   it("parses beat lists with note-value offsets", () => {
     expect(parser.parse("1|1,2+n/16,2+n/8,2+n3/16")).toStrictEqual([
       { bar: 1, beat: 1 },
@@ -90,6 +105,12 @@ describe("BarBeatScript Parser - time declarations", () => {
   it("parses repeat pattern with note-value offset start (positions still meter-relative)", () => {
     expect(parser.parse("1|2+n/12x3@n1/3")).toStrictEqual([
       { bar: 1, beat: { start: obeat(2, 1, 12), times: 3, step: 1 / 3 } },
+    ]);
+  });
+
+  it("parses repeat pattern with a decimal-base offset start (1|1.5+n/12x3)", () => {
+    expect(parser.parse("1|1.5+n/12x3@n1/3")).toStrictEqual([
+      { bar: 1, beat: { start: obeat(1.5, 1, 12), times: 3, step: 1 / 3 } },
     ]);
   });
