+++ src/notation/transform/tests/parser/transform-parser.test.ts
@@ -183,85 +183,6 @@ describe("Transform Parser", () => {
     });
   });
 
-  describe("time range selectors", () => {
-    it("parses bar|beat-bar|beat range", () => {
-      const result = parser.parse("1|1-3|1: velocity += 10");
-
-      expect(result[0]!.timeRange).toStrictEqual({
-        startBar: 1,
-        startBeat: 1,
-        endBar: 3,
-        endBeat: 1,
-      });
-    });
-
-    it("parses fractional beats in range", () => {
-      const result = parser.parse("1|1.5-2|3.5: velocity += 10");
-
-      expect(result[0]!.timeRange).toStrictEqual({
-        startBar: 1,
-        startBeat: 1.5,
-        endBar: 2,
-        endBeat: 3.5,
-      });
-    });
-
-    it("parses +n note-value offset bounds", () => {
-      const result = parser.parse("1|1+n/12-2|2-n/24: velocity += 10");
-
-      // 4/4 default: n/12 = 1/3 beat, n/24 = 1/6 beat
-      expect(result[0]!.timeRange!.startBeat).toBeCloseTo(1.3333);
-      expect(result[0]!.timeRange!.endBeat).toBeCloseTo(1.8333);
-    });
-
-    it("resolves n offsets meter-relative to timeSigDenominator", () => {
-      // n/12 = 1/12 whole note → 1/3 beat in 4/4, 2/3 beat in 6/8.
-      const in44 = parser.parse("1|1+n/12-2|1: velocity += 10", {
-        timeSigDenominator: 4,
-      });
-      const in68 = parser.parse("1|1+n/12-2|1: velocity += 10", {
-        timeSigDenominator: 8,
-      });
-
-      expect(in44[0]!.timeRange!.startBeat).toBeCloseTo(1.3333);
-      expect(in68[0]!.timeRange!.startBeat).toBeCloseTo(1.6667);
-    });
-
-    it("rejects bare fractions in a range bound", () => {
-      expect(() => parser.parse("1|4/3-2|1: velocity += 10")).toThrow(
-        /bare fraction/,
-      );
-    });
-
-    it("rejects mixed numbers in a range bound", () => {
-      expect(() => parser.parse("1|1+1/3-2|1: velocity += 10")).toThrow(
-        /note-value form/,
-      );
-    });
-
-    it("borrows across the bar line for a -n bound before the downbeat", () => {
-      // `2|1-n/12` start bound = just before the bar-2 downbeat → bar 1, beat 4⅔
-      // in 4/4. The separator `-` before `3|1` is not consumed as an offset.
-      const result = parser.parse("2|1-n/12-3|1: velocity += 10");
-
-      expect(result[0]!.timeRange!.startBar).toBe(1);
-      expect(result[0]!.timeRange!.startBeat).toBeCloseTo(4.6667);
-      expect(result[0]!.timeRange!.endBar).toBe(3);
-      expect(result[0]!.timeRange!.endBeat).toBe(1);
-    });
-
-    it("resolves a -n bound before 1|1 to a pre-clip-start position (no throw)", () => {
-      // `1|1-n/2` = bar 1 minus a half note → bar 0, beat 3 (negative time when
-      // resolved). Allowed, not rejected.
-      const result = parser.parse("1|1-n/2-2|1: velocity += 10");
-
-      expect(result[0]!.timeRange!.startBar).toBe(0);
-      expect(result[0]!.timeRange!.startBeat).toBe(3);
-      expect(result[0]!.timeRange!.endBar).toBe(2);
-      expect(result[0]!.timeRange!.endBeat).toBe(1);
-    });
-  });
-
   describe("combined selectors", () => {
     it("parses pitch with time range", () => {
       const result = parser.parse("E3 1|1-2|1: velocity += 10");
